fix type issue

+++ src/main/java/de/wi2020sebgroup1/cinema/model/User.java
@@ -35,10 +35,10 @@ public class User {
 		this.lastName = lastName;
 		this.email = email;
 	}
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getFirstName() {
